- Move (and slightly clean up) RuleScheduleStateController... Still needs so much love. - Removed unused code (RuleDeviceSection)
