--- cornea/src/main/java/arcus/cornea/RuleDeviceSection.java
@@ -1,50 +1,0 @@
-/*
- *  Copyright 2019 Arcus Project.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package arcus.cornea;
-
-import com.iris.client.model.RuleModel;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class RuleDeviceSection {
-
-    private List<RuleModel> rules;
-
-
-    public RuleDeviceSection() {
-
-
-    }
-
-    private boolean isRulesNull(){
-        return rules == null;
-    }
-
-    public void addRule(RuleModel ruleModel){
-        if (isRulesNull()){
-            rules = new ArrayList<RuleModel>();
-        }
-        rules.add(ruleModel);
-    }
-
-
-
-    public List<RuleModel> getRules() {
-        return rules;
-    }
-
-}
